rename: home directory 구조 변경 #18
